Refactor: 모임통장 - SNS 기능 코드 분리
