+++ src/main/java/team/hanaro/hanamate/domain/moimWallet/MoimWalletService.java
@@ -10,6 +10,7 @@ import team.hanaro.hanamate.domain.MyWallet.Repository.TransactionRepository;
 import team.hanaro.hanamate.domain.User.Repository.UsersRepository;
 import team.hanaro.hanamate.domain.moimWallet.dto.MoimWalletRequestDto;
 import team.hanaro.hanamate.domain.moimWallet.dto.MoimWalletResponseDto;
+import team.hanaro.hanamate.domain.moimWallet.dto.SNSResponseDTO;
 import team.hanaro.hanamate.domain.moimWallet.repository.ArticleRepository;
 import team.hanaro.hanamate.domain.moimWallet.repository.MoimWalletRepository;
 import team.hanaro.hanamate.domain.moimWallet.repository.MoimWalletAndUserRepository;
@@ -33,6 +34,8 @@ public class MoimWalletService {
     private final TransactionRepository transactionRepository;
     private final ArticleRepository articleRepository;
     private final Response response;
+
+    private final SNSService snsService;
 
     public ResponseEntity<?> getMoimWalletListByLoginId(MoimWalletRequestDto.findAllMoimWalletDTO moimWalletDto) {
         Optional<User> byLoginId = usersRepository.findByLoginId(moimWalletDto.getUserId());
@@ -98,15 +101,7 @@ public class MoimWalletService {
                     .map(transaction -> {
                         MoimWalletResponseDto.MoimWalletTransactionsDTO moimTransactionResponseDTO = new MoimWalletResponseDto.MoimWalletTransactionsDTO(transaction);
                         Optional<Article> optionalArticle = Optional.ofNullable(transaction.getArticle());
-                        optionalArticle.map(article -> {
-                            MoimWalletResponseDto.ArticleResponseDTO dto = new MoimWalletResponseDto.ArticleResponseDTO(article);
-                            List<MoimWalletResponseDto.CommentResponseDTO> commentResponseDTOList = article.getComments()
-                                    .stream()
-                                    .map(MoimWalletResponseDto.CommentResponseDTO::new)
-                                    .collect(Collectors.toList());
-                            dto.setCommentList(commentResponseDTOList);
-                            return dto;
-                        }).ifPresent(moimTransactionResponseDTO::setArticle);
+                        snsService.getArticleResponseDTO(optionalArticle).ifPresent(moimTransactionResponseDTO::setArticle);
                         return moimTransactionResponseDTO;
                     }).collect(Collectors.toList());
             MoimWalletResponseDto.MoimWalletDTO moimWalletDTO = new MoimWalletResponseDto.MoimWalletDTO(savedMoimWallet.get());
@@ -114,6 +109,8 @@ public class MoimWalletService {
             return response.success(moimWalletDTO, "성공", HttpStatus.OK);
         }
     }
+
+
 
     public ResponseEntity<?> updateMoimWalletInfo(MoimWalletRequestDto.UpdateMoimWalletInfoRequestDTO updateMoimWalletDTO) {
         Optional<MoimWallet> optionalMoimWallet = moimWalletRepository.findById(updateMoimWalletDTO.getMoimWalletId());
@@ -153,55 +150,5 @@ public class MoimWalletService {
         return response.success("제작중");
     }
 
-    public ResponseEntity<?> writeArticle(MoimWalletRequestDto.WriteArticleRequestDTO articleDTO) throws SQLException, IOException {
-        //1. 트랜잭션 id로 해당 트랜잭션 찾아온다.
-        Optional<Transactions> savedTransaction = transactionRepository.findById(articleDTO.getTransactionId());
-        if (savedTransaction.isEmpty()) {
-            return response.fail("잘못된 거래내역 입니다.", HttpStatus.BAD_REQUEST);
-        }
-        MoimWalletResponseDto.ArticleResponseDTO articleResponseDTO = null;
-        if (savedTransaction.get().getArticle() == null) {
-            //존재하지 않으면 생성
-            Article article = Article.builder()
-                    .content(articleDTO.getContent())
-                    .title(articleDTO.getTitle())
-                    .imageId(articleDTO.getImage())
-                    .transaction(savedTransaction.get()).build();
-            articleRepository.save(article);
-            articleResponseDTO = new MoimWalletResponseDto.ArticleResponseDTO(article);
-            return response.success(articleResponseDTO, "성공적으로 글이 작성되었습니다.", HttpStatus.OK);
-        } else {
-            //존재하면 수정
-            Article savedArticle = savedTransaction.get().getArticle();
-            savedArticle.setContent(articleDTO.getContent());
-            savedArticle.setTitle(articleDTO.getTitle());
-            savedArticle.setImageId(articleDTO.getImage());
-            articleRepository.save(savedArticle);
-            articleResponseDTO = new MoimWalletResponseDto.ArticleResponseDTO(savedArticle);
-            List<MoimWalletResponseDto.CommentResponseDTO> commentDTOList = new ArrayList<>();
-            for (Comment comment : savedArticle.getComments()) {
-                MoimWalletResponseDto.CommentResponseDTO temp = new MoimWalletResponseDto.CommentResponseDTO(comment);
-                commentDTOList.add(temp);
-            }
-            articleResponseDTO.setCommentList(commentDTOList);
-            return response.success(articleResponseDTO, "글이 수정되었습니다.", HttpStatus.OK);
-        }
-        //2. 받아온 작성글(Article)을 DTO에서 새 엔티티로 반환
-    }
 
-
-    public ResponseEntity<?> deleteArticle(MoimWalletRequestDto.DeleteRequestDTO deleteRequestDTO) {
-        Optional<Article> savedArticle = articleRepository.findById(deleteRequestDTO.getRequestId());
-        if (savedArticle.isEmpty()) {
-            return response.fail("존재하지 않는 글입니다.", HttpStatus.BAD_REQUEST);
-        } else {
-            articleRepository.delete(savedArticle.get());
-            articleRepository.flush();
-            if (articleRepository.findById(deleteRequestDTO.getRequestId()).isEmpty()) {
-                return response.success(deleteRequestDTO.getRequestId() + "글이 성공적으로 삭제되었습니다.");
-            } else {
-                return response.fail("글 삭제에 실패 했습니다.", HttpStatus.BAD_REQUEST);
-            }
-        }
-    }
 }
